refactor: split gallery module to render and api parts
split gallery module to render and api parts

MODULE-1-PART-3-gallery
